+++ packages/provider/provider.js
@@ -1,6 +1,3 @@
 import connect from "./connect";
-import AuthorProfileProvider from "./author-profile-provider";
 
 export default connect;
-
-export { AuthorProfileProvider };
